requirement.txt
streamlit
duckduckgo-search
python-dotenv
openai
groq
